Feature/get data from api

+++ app/src/main/java/com/example/cookup/models/Ingredient.kt
@@ -8,18 +8,18 @@ import kotlinx.serialization.Serializable
 @Serializable
 @Parcelize
 data class Ingredient(
-    val quantity: Int,
+    val quantity: Double,
     val unit: EIngredientUnit?,
     val name: String
 ) : Parcelable{
 
-    constructor() : this(0, null, "")
+    constructor() : this(0.0, null, "")
 
     companion object {
         fun fromMap(map: Map<String, Any>): Ingredient {
             return Ingredient(
                 name = map["name"] as? String ?: "",
-                quantity = (map["quantity"] as? Number)?.toInt() ?: 0,
+                quantity = (map["quantity"] as? Number)?.toDouble() ?: 0.0,
                 unit = map["unit"] as? EIngredientUnit ?: null
             )
         }
